simplify

+++ kernel/trap.c
@@ -33,6 +33,7 @@ trapinithart(void)
 // handle an interrupt, exception, or system call from user space.
 // called from trampoline.S
 //
+
 void
 usertrap(void)
 {
@@ -70,22 +71,19 @@ usertrap(void)
   } else {
     uint64 scause = r_scause();
     uint64 stval = r_stval();
-    if ((scause == 13 || scause == 15) && stval <= p->sz) { // page fault
+    if ((scause == 13 || scause == 15) &&
+         stval < p->sz &&
+         PGROUNDUP(stval) != PGROUNDDOWN(p->trapframe->sp)) { // page fault
       uint64 va = PGROUNDDOWN(stval);
-      pte_t *pte = walk(p->pagetable, va, 0);
-      if (pte != 0 && (*pte & PTE_V) == 1 && (*pte & PTE_U) == 0) {
-        p->killed = 1; // stack overflow
+      char *mem = kalloc();
+      if(mem == 0){
+        p->killed = 1; // kalloc fails, kill process
       } else {
-        char *mem = kalloc();
-        if(mem == 0){
-          p->killed = 1; // kalloc fails, kill process
-        } else {
-          memset(mem, 0, PGSIZE);
-          if(mappages(p->pagetable, va, PGSIZE, (uint64)mem, PTE_W|PTE_X|PTE_R|PTE_U) != 0){
-            kfree(mem);
-            uvmdealloc(p->pagetable, va, va);
-            panic("can't map");
-          }
+        memset(mem, 0, PGSIZE);
+        if(mappages(p->pagetable, va, PGSIZE, (uint64)mem, PTE_W|PTE_X|PTE_R|PTE_U) != 0){
+          kfree(mem);
+          uvmdealloc(p->pagetable, va, va);
+          panic("can't map");
         }
       }
     } else {
